+++ modules/core/hardware.nix
@@ -5,8 +5,6 @@
       extraBackends = [pkgs.sane-airscan];
       disabledDefaultBackends = ["escl"];
     };
-    logitech.wireless.enable = false;
-    logitech.wireless.enableGraphical = false;
     graphics.enable = true;
     enableRedistributableFirmware = true;
     keyboard.qmk.enable = true;