+++ CHANGELOG.md
@@ -17,6 +17,9 @@
       ! Couldn't find npm. `lsp-mode' needs npm to auto-install some LSP servers. For more information, see https://emacs-lsp.github.io/lsp-mode/page/languages/.
   ```
 
+- Added:
+  - Overly for dwarfs build error
+
 - Removed:
   - overlays for glaze and tumbler
     - `overlay.nix` is now empty template